+++ mkrtos_user/lib/mlibc/src/internal/pthread_impl.h
@@ -65,7 +65,7 @@ struct pthread {
 	uintptr_t canary;
 	uintptr_t *dtv;
 #endif
-	int ctid;
+	unsigned long ctid;
 	unsigned long hd;
 };
